Move some data and Splitting (#32)
* move friend area settings to src

* split out playtime into different C file and decomp a few more funcs

* time -> play_time

* fix include

* split out more wonder mail stuff

* unify 203B2C0 and unkDungeon structs. snuck in a decomp func too

* split out more menu screens

* work some more on trade items

* doc a few more things

* add a close nonmatch and decomp nullsub and small func

* rollback struct attempt since it causes nonmatchings

* only one func left in asm

* death to trade_item asm and unify into one C file

* more docing and cleanup

* label main menu and trade items menu global menu struct

* doc itemMode

+++ include/friend_area.h
@@ -4,11 +4,7 @@
 struct FriendAreaSettings
 {
     s16 num_pokemon;
-    u8 unlock_condition;
-    // 0 - Shop (Story)
-    // 1 - Shop (Story Post-game)
-    // 2 - Wonder Mail
-    // 3 - Legendary Request (unlocks when lengendary joins)
+    u16 unlock_condition;
     u32 price;
 };
